Move FileId to parser module
